requirements.txt
accelerate==1.7.0
aiohttp==3.12.14
beautifulsoup4==4.13.4
braceexpand==0.1.7
clip_retrieval==2.44.0
compel==2.1.1
consisid_eva_clip==1.0.2
controlnet_aux==0.0.10
d4rl==1.1
datasets==4.0.0
decord==0.6.0
deepspeed==0.17.2
diffusion==6.11.6
easydict==1.13
easyocr==1.7.2
einops==0.8.1
facexlib==0.3.0
faiss==1.5.3
fastapi==0.116.1
flash_attn==2.8.1
flax==0.10.7
ftfy==6.3.1
GitPython==3.1.44
gpustat==1.1.1
gym==0.26.2
huggingface_hub==0.31.4
imageio==2.37.0
importlib_metadata==8.0.0
importlib_metadata==8.7.0
imwatermark==0.0.2
insightface==0.7.3
intel_extension_for_pytorch==2.7.0
jax==0.6.2
jaxlib==0.6.2
k_diffusion==0.1.1.post1
kornia==0.8.1
librosa==0.11.0
ligo==1.0.11
lpips==0.1.4
matplotlib==3.10.3
msgpack_python==0.5.6
muse==0.1
neural_compressor==3.4.1
note_seq==0.0.5
numpy==2.3.1
onnx==1.18.0
onnx_graphsurgeon==0.5.8
opencv_python_headless==4.12.0.88
optax==0.2.5
optimum==1.26.1
packaging==25.0
pandas==2.3.1
parameterized==0.9.0
Pillow==11.3.0
polygraphy==0.49.26
prodigyopt==1.1.2
psutil==7.0.0
psycopg2==2.9.10
pycuda==2025.1.1
pydantic==2.11.7
PyGithub==2.6.1
pytest==8.4.1
PyYAML==6.0.2
PyYAML==6.0.2
ram==0.1
regex==2024.11.6
Requests==2.32.4
requests_mock==1.12.1
safetensors==0.5.3
sageattention==1.0.6
scipy==1.16.0
sentencepiece==0.2.0
setuptools==80.7.1
skimage==0.0
slack_sdk==3.36.0
structlog==25.4.0
t5x==0.0.0
tabulate==0.9.0
tensorrt==10.12.0.36
termcolor==3.1.0
timm==1.0.17
torch==2.7.1
torch_npu==2.5.1
torch_xla==2.7.0
torchprofile==0.0.4
torchsde==0.2.6
torchvision==0.22.1
tqdm==4.67.1
typing_extensions==4.14.1
uvicorn==0.35.0
wandb==0.21.0
webdataset==1.0.2
xformers==0.0.31.post1
